Create LinkedList : addAtBeginning
Insertion in singly linked list at the beginning

C Program for insertion at the beginning of the Singly Linked List. A Singly linked list is made up of many nodes which are connected. Every node is mainly divided into two parts, one part holds the data and the other part is the link that connects to the next node.